CMakeLists.txt improved. Use epoll_event.u64 to replace epoll_event.fd. No longer need to use GetPointer. Forget initialize configuration. Remove cache for native DNS query method. Add mutex for memory pool. Variable name more concisely. OnWritable bug fix. Other improved.

+++ SimpleProxy/dispatcher/ibusiness_event.hh
@@ -1,15 +1,17 @@
 #ifndef IBUSINESS_EVENT_HEADER
 #define IBUSINESS_EVENT_HEADER
 
+#include <cstdint>
+
 class IPoller;
 
 class IBusinessEvent {
 public:
     virtual ~IBusinessEvent() {}
-    virtual void OnAcceptable(int){};
-    virtual void OnCloseable(int){};
-    virtual void OnReadable(int){};
-    virtual void OnWritable(int){};
+    virtual void OnAcceptable(uintptr_t){};
+    virtual void OnCloseable(uintptr_t){};
+    virtual void OnReadable(uintptr_t){};
+    virtual void OnWritable(uintptr_t){};
     IPoller* poller_ = nullptr;
 };
